Add "Move" file command
This commands presents the full path of the file to be moved, allowing
the user to type the destination and renaming it in just one operation.

If an error with the move operation occurs, a temporary message will be
shown in the status bar of the current view, and it will be discarded
after 5 seconds.