"New file" command
This commands allows file creation anywhere in the filesystem and it
will attempt to create all the subtrees needed to realize the path.

The command will attempt to check if the file exists already and exit
early if that's the case; otherwise no other checks are performed: if
the operation fails for any reason, a `status_message` is emitted with
the contents of the error/exception.