fix: nonetype error when creating new folder
The new command uses the active view file name as the path from which to
construct the new folder, but when the view is empty there is no real
file on disk and the file name is None.

The `os.path` functions like `isdir` and `exists` raise a NoneType
exception when given None. This patch adds a NoneType check.

Close #55
Re #56